requirements.txt
google-api-python-client
google-auth
google-auth-oauthlib
nltk
scikit-learn
beautifulsoup4